Refactor login activities upload logic into ActivitiesRepository
Moves the chunked, concurrent login activities API upload loop out of `UploadManager` and fully encapsulates it within `ActivitiesRepositoryImpl`. Injects `ApiInterface` into `ActivitiesRepositoryImpl` via Hilt.
