+++ app/src/main/java/org/ole/planet/myplanet/repository/ActivitiesRepository.kt
@@ -25,4 +25,5 @@ interface ActivitiesRepository {
     suspend fun getRecentLogin(): RealmOfflineActivity?
     fun serializeLoginActivities(activity: RealmOfflineActivity, context: android.content.Context): JsonObject
     fun bulkInsertLoginActivitiesFromSync(realm: io.realm.Realm, jsonArray: com.google.gson.JsonArray)
+    suspend fun uploadActivities()
 }
