+++ app/src/main/java/org/ole/planet/myplanet/repository/ActivitiesRepositoryImpl.kt
@@ -15,6 +15,14 @@ import org.ole.planet.myplanet.model.RealmCourseActivity
 import org.ole.planet.myplanet.model.RealmOfflineActivity
 import org.ole.planet.myplanet.model.RealmRemovedLog
 import org.ole.planet.myplanet.model.RealmResourceActivity
+import kotlinx.coroutines.async
+import kotlinx.coroutines.awaitAll
+import kotlinx.coroutines.coroutineScope
+import kotlinx.coroutines.sync.Semaphore
+import kotlinx.coroutines.sync.withPermit
+import org.ole.planet.myplanet.data.api.ApiInterface
+import org.ole.planet.myplanet.utils.UrlUtils
+import android.util.Log
 import org.ole.planet.myplanet.model.RealmTeamLog
 import org.ole.planet.myplanet.repository.TeamsRepository
 import org.ole.planet.myplanet.services.UserSessionManager
@@ -25,7 +33,8 @@ class ActivitiesRepositoryImpl @Inject constructor(
     @RealmDispatcher realmDispatcher: CoroutineDispatcher,
     @ApplicationContext private val context: Context,
     private val teamsRepository: Lazy<TeamsRepository>,
-    private val userRepository: Lazy<UserRepository>
+    private val userRepository: Lazy<UserRepository>,
+    private val apiInterface: ApiInterface
 ) : RealmRepository(databaseService, realmDispatcher), ActivitiesRepository {
     override suspend fun getOfflineActivities(userName: String, type: String): List<RealmOfflineActivity> {
         return queryList(RealmOfflineActivity::class.java) {
@@ -303,6 +312,42 @@ class ActivitiesRepositoryImpl @Inject constructor(
             ob.addProperty("_rev", activity._rev)
         }
         return ob
+    }
+
+        override suspend fun uploadActivities() {
+        val activitiesToUpload = getUnuploadedLoginActivities()
+
+        activitiesToUpload.chunked(50).forEach { batch ->
+            val successfulUpdates = mutableMapOf<String, com.google.gson.JsonObject?>()
+
+            val semaphore = Semaphore(6)
+            coroutineScope {
+                val deferreds = batch.map { activityData ->
+                    async {
+                        try {
+                            val `object` = semaphore.withPermit {
+                                apiInterface.postDoc(
+                                    UrlUtils.header, "application/json",
+                                    "${UrlUtils.getUrl()}/login_activities", activityData.serialized
+                                ).body()
+                            }
+                            activityData.id to `object`
+                        } catch (e: java.io.IOException) {
+                            Log.e("ActivitiesRepository", "Exception in UploadManager", e)
+                            null
+                        }
+                    }
+                }
+                deferreds.awaitAll().filterNotNull().forEach { (id, obj) ->
+                    successfulUpdates[id] = obj
+                }
+            }
+
+            if (successfulUpdates.isNotEmpty()) {
+                val idsToUpdate = successfulUpdates.keys.toTypedArray()
+                markActivitiesUploaded(idsToUpdate, successfulUpdates)
+            }
+        }
     }
 
     override fun bulkInsertLoginActivitiesFromSync(realm: io.realm.Realm, jsonArray: com.google.gson.JsonArray) {
