+++ app/src/main/java/org/ole/planet/myplanet/services/UploadManager.kt
@@ -456,39 +456,7 @@ class UploadManager @Inject constructor(
         }
 
         try {
-            val activitiesToUpload = activitiesRepository.getUnuploadedLoginActivities()
-
-            activitiesToUpload.chunked(BATCH_SIZE).forEach { batch ->
-                val successfulUpdates = mutableMapOf<String, com.google.gson.JsonObject?>()
-
-                val semaphore = Semaphore(6)
-                coroutineScope {
-                    val deferreds = batch.map { activityData ->
-                        async {
-                            try {
-                                val `object` = semaphore.withPermit {
-                                    apiInterface.postDoc(
-                                        UrlUtils.header, "application/json",
-                                        "${UrlUtils.getUrl()}/login_activities", activityData.serialized
-                                    ).body()
-                                }
-                                activityData.id to `object`
-                            } catch (e: java.io.IOException) {
-                                Log.e(TAG, "Exception in UploadManager", e)
-                                null
-                            }
-                        }
-                    }
-                    deferreds.awaitAll().filterNotNull().forEach { (id, obj) ->
-                        successfulUpdates[id] = obj
-                    }
-                }
-
-                if (successfulUpdates.isNotEmpty()) {
-                    val idsToUpdate = successfulUpdates.keys.toTypedArray()
-                    activitiesRepository.markActivitiesUploaded(idsToUpdate, successfulUpdates)
-                }
-            }
+            activitiesRepository.uploadActivities()
 
             uploadTeamActivities()
